+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ Unreleased
 Added
 -----
 
+- **Projectroles**
+    - Helper ``email.send_generic_mail()`` (#280)
 - **Samplesheets**
     - Display original study/assay filenames as tooltips (#283)
     - Display assays for samples in search results (#157)
@@ -24,6 +26,8 @@ Changed
 
 - **Irodsbackend**
     - Optimized iRODS queries for increased performance (#242)
+- **Projectroles**
+    - Minor email refactoring (#280)
 - **Samplesheets**
     - Search for VCF files under all family members in germline app (#275)
     - Include ``alt_name`` in GenericMaterial search (#285)